+++ util/import/getODataURL.m
@@ -1,8 +1,8 @@
-function endpoint = getODataURL(cellNum, source, urlType)
+function endpoint = getODataURL(ID, source, urlType)
 	% GETODATAURL  Returns OData URL string
 	%
 	% Inputs:
-	%	cellNum 		Cell ID number
+	%	ID              Unique ID number
 	%	source 			'inferior', 'temporal' ('i', 't')
 	% 	urlType 		'neuron', 'location', 'link', 'child', 'scale'
 	% Use:
@@ -14,14 +14,14 @@ function endpoint = getODataURL(cellNum, source, urlType)
 
 	switch urlType
 		case 'neuron'
-			endpoint = [baseURL 'Structures(' num2str(cellNum) ')/'];
+			endpoint = [baseURL 'Structures(' num2str(ID) ')/'];
 		case 'location'
-			endpoint = [baseURL 'Structures(' num2str(cellNum), ')/Locations/'];
+			endpoint = [baseURL 'Structures(' num2str(ID), ')/Locations/'];
 		case 'link'
-			endpoint = [baseURL 'Structures(' num2str(cellNum)... 
+			endpoint = [baseURL 'Structures(' num2str(ID)... 
 				')/LocationLinks/?$select=A,B'];
 		case 'child'
-			endpoint = [baseURL 'Structures(' num2str(cellNum) ')/Children',...
+			endpoint = [baseURL 'Structures(' num2str(ID) ')/Children',...
                 '?$select=ID,TypeID,Tags,ParentID,Label'];
 		case 'scale'
 			endpoint = [baseURL 'Scale'];